+++ instructor/lib/instructor.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
 require "active_support"
+require "active_support/inflector"
 require "active_model"
 
 require "short_circu_it"
+require "tablesalt"
 require "technologic"
 
 require "instructor/version"