+++ instructor/spec/instructor/base_spec.rb
@@ -13,5 +13,4 @@ RSpec.describe Instructor::Base, type: :instructor do
   it { is_expected.to include_module Instructor::Arguments }
   it { is_expected.to include_module Instructor::Options }
   it { is_expected.to include_module Instructor::Core }
-  it { is_expected.to include_module Instructor::String }
 end