+++ lib/spicerack.rb
@@ -9,4 +9,7 @@ require "short_circu_it"
 require "technologic"
 require "tablesalt"
 
+require "spicerack/array_index"
+require "spicerack/hash_model"
+
 module Spicerack; end