+++ lib/spicerack/spec_helper.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require "rspice"
+require "instructor/spec_helper"
+
+require_relative "custom_matchers"