+++ spec/integration/task_hash_model_spec.rb
@@ -1,13 +1,15 @@
 # frozen_string_literal: true
 
 RSpec.describe TaskHashModel, type: :integration do
-  subject(:hash_model) { described_class.for(hash) }
+  subject(:hash_model) { described_class.for(data) }
 
-  let(:hash) do
+  let(:data) do
     {}
   end
 
-  it { is_expected.to inherit_from Tablesalt::HashModel }
+  it { is_expected.to inherit_from Spicerack::HashModel }
+  it { is_expected.to define_field :started_at, :datetime }
+  it { is_expected.to define_field :finished_at, :datetime }
 
   describe "._fields" do
     subject { hash_model._fields }