+++ spec/support/test_classes/hash_model.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 class ExampleHashModelBase
-  include Tablesalt::HashModel
+  include Spicerack::HashModel
 
   class << self
-    def for(hash)
-      new.tap { |model| model.hash = hash }
+    def for(data)
+      new.tap { |model| model.data = data }
     end
   end
 end