+++ tablesalt/lib/tablesalt.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 require "active_support"
+require "short_circu_it"
 
 require "tablesalt/version"
 
-require "tablesalt/array_index"
-require "tablesalt/hash_model"
+require "tablesalt/stringable_object"
 
 module Tablesalt; end