+++ tablesalt/spec/spec_helper.rb
@@ -2,10 +2,4 @@
 
 require_relative "../../shared/spec_helper"
 
-require_relative "support/shared_examples/a_hash_model_reader"
-require_relative "support/shared_examples/a_hash_model_writer"
-require_relative "support/shared_examples/a_task_hash_model"
-
 require "tablesalt"
-
-require_relative "support/test_classes/hash_model"