+++ redis_hash/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.9.6
+
+*Release Date*: 5/21/2019
+
+- Added `piplined`, `multi/exec`, and `persist` methods
+- Added `schema` for enforcement of a whitelist on redis hash keys
+
 ## v0.9.5
 
 *Release Date*: 5/20/2019