+++ tablesalt/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.9.6
+
+*Release Date*: 5/21/2019
+
+- Move `ArrayIndex` to spicerack gem
+- Move `HashModel` to spicerack gem
+- Move `StringableObject` to tablesalt gem
+
 ## v0.9.5
 
 *Release Date*: 5/20/2019